lib: Implement converge_parents
The implementation is fairly simple. The idea is to pick one of  the
divergent commits, and then use its parent(s) as the parent(s) of the
"solution". This is trivial when all divergent commits have exactly the
same parents.

The only tricky part is that some commits cannot be used as parents.
This is easiest to explain with an example: say D1, D2 and D3 are the
divergent commits. If D3 is a descendant of D1 (in the commit graph,
NOT in the sense of evolution) then we cannot use D3's parents for the
solution because we cannot rebase D1 on top of that.